+++ bauh/gems/arch/pacman.py
@@ -141,7 +141,10 @@ def install_as_process(pkgpaths: Iterable[str], root_password: str, file: bool, 
     else:
         cmd = ['pacman', '-S', *pkgpaths, '--noconfirm']  # pkgpath = pkgname
 
-    return SimpleProcess(cmd=cmd, root_password=root_password, cwd=pkgdir)
+    return SimpleProcess(cmd=cmd,
+                         root_password=root_password,
+                         cwd=pkgdir,
+                         error_phrases={"error: failed to prepare transaction"})
 
 
 def list_desktop_entries(pkgnames: Set[str]) -> List[str]:
@@ -692,7 +695,9 @@ def list_installed_names() -> Set[str]:
 
 
 def upgrade_several(pkgnames: Iterable[str], root_password: str) -> SimpleProcess:
-    return SimpleProcess(cmd=['pacman', '-S', *pkgnames, '--noconfirm'], root_password=root_password)
+    return SimpleProcess(cmd=['pacman', '-S', *pkgnames, '--noconfirm'],
+                         root_password=root_password,
+                         error_phrases={'error: failed to prepare transaction'})
 
 
 def remove_several(pkgnames: Iterable[str], root_password: str) -> SystemProcess: