+++ CHANGELOG.md
@@ -10,7 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixes
 - Arch
+    - not stopping the upgrade process if a transaction error happens
     - wrong dialog titles
+    
 
 ## [0.9.0] - 2020-04-15
 ### Features